Update internal SDL library from 2.0.12 to 2.0.16
Will probably result in a new Extension version, but as there's extra changes on the Fusion code side, I'll leave this honking update in a separate commit.

Some edits to SDL update README, as they fixed some things about the process and made it simpler. I also made the change list alphabetical so it's easier to match by eye.